consumer/requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
idna==3.7
Jinja2==3.1.4
kafka-python==2.0.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
mysql-connector-python==8.4.0
numpy==1.21.0
packaging==24.1
preshed==3.0.9
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
smart-open==7.0.4
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.4
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
